requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
defusedxml==0.7.1
Django==5.0.2
django-allauth==0.61.1
django-chartjs==2.3.0
django-crispy-forms==2.1
django-embed-video==1.4.9
django-extensions==3.2.3
exchange_calendars==4.5.5
fonttools==4.53.1
html5lib==1.1
idna==3.8
kaleido==0.2.1
kiwisolver==1.4.6
korean-lunar-calendar==0.3.1
lxml==5.3.0
matplotlib==3.9.2
matplotlib-inline==0.1.6
mechanize==0.4.7
mindee==4.10.0
monotonic==1.6
numpy==1.26.4
oauthlib==3.2.2
packaging==24.1
pandas==2.2.1
pandas-datareader==0.10.0
pandas_market_calendars==4.4.0
pillow==10.4.0
plotly==5.24.0
pycparser==2.22
PyJWT==2.9.0
pyluach==2.2.0
pyparsing==3.1.4
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
six==1.16.0
sqlparse==0.5.1
tenacity==9.0.0
toolz==0.12.1
traitlets==5.14.3
tzdata==2024.1
urllib3==2.2.2
webencodings==0.5.1
